refactor(apps/metrics): move to apis/web/metrics
