+++ src/dashboard/apigateway/apigateway/apis/web/metrics/__init__.py
@@ -15,27 +15,3 @@
 # We undertake not to change the open source license (MIT license) applicable
 # to the current version of the project delivered to anyone in the future.
 #
-import pytest
-
-from apigateway.apps.metrics.utils import MetricsSmartTimeRange
-
-
-class TestMetricsSmartTimeRange:
-    @pytest.mark.parametrize(
-        "time_range_minutes, expected",
-        [
-            (10, "1m"),
-            (59, "1m"),
-            (60, "1m"),
-            (300, "5m"),
-            (360, "5m"),
-            (720, "10m"),
-            (1440, "30m"),
-            (4320, "1h"),
-            (10080, "3h"),
-            (20000, "12h"),
-        ],
-    )
-    def test_get_recommended_step(self, time_range_minutes, expected):
-        smart_time_range = MetricsSmartTimeRange(time_range=time_range_minutes * 60)
-        assert smart_time_range.get_recommended_step() == expected
