+++ src/dashboard/apigateway/apigateway/apis/web/metrics/views.py
@@ -1,3 +1,4 @@
+# -*- coding: utf-8 -*-
 #
 # TencentBlueKing is pleased to support the open source community by making
 # 蓝鲸智云 - API 网关(BlueKing - APIGateway) available.
@@ -17,7 +18,17 @@
 #
 import math
 
+from django.http import Http404
+from drf_yasg.utils import swagger_auto_schema
+from rest_framework import generics, status
+
+from apigateway.apps.metrics.constants import DimensionEnum, MetricsEnum
+from apigateway.apps.metrics.prometheus.dimension import DimensionMetricsFactory
+from apigateway.core.models import Resource, Stage
+from apigateway.utils.responses import OKJsonResponse
 from apigateway.utils.time import SmartTimeRange
+
+from .serializers import MetricsQueryInputSLZ
 
 
 class MetricsSmartTimeRange(SmartTimeRange):
@@ -61,3 +72,46 @@ class MetricsSmartTimeRange(SmartTimeRange):
             index = 6
 
         return step_options[index]
+
+
+class QueryRangeApi(generics.ListAPIView):
+    @swagger_auto_schema(
+        query_serializer=MetricsQueryInputSLZ,
+        responses={status.HTTP_200_OK: ""},
+        tags=["Metrics"],
+    )
+    def get(self, request, *args, **kwargs):
+        slz = MetricsQueryInputSLZ(data=request.query_params)
+        slz.is_valid(raise_exception=True)
+
+        data = slz.validated_data
+
+        stage_name = Stage.objects.get_name(request.gateway.id, data["stage_id"])
+        if not stage_name:
+            raise Http404
+
+        resource_name = None
+        if data.get("resource_id"):
+            resource_name = Resource.objects.get_name(request.gateway.id, data["resource_id"])
+
+        smart_time_range = MetricsSmartTimeRange(
+            data.get("time_start"),
+            data.get("time_end"),
+            data.get("time_range"),
+        )
+        time_start, time_end = smart_time_range.get_head_and_tail()
+        step = smart_time_range.get_recommended_step()
+
+        metrics = DimensionMetricsFactory.create_dimension_metrics(
+            DimensionEnum(data["dimension"]), MetricsEnum(data["metrics"])
+        )
+        data = metrics.query_range(
+            gateway_name=request.gateway.name,
+            stage_name=stage_name,
+            resource_name=resource_name,
+            start=time_start,
+            end=time_end,
+            step=step,
+        )
+
+        return OKJsonResponse("OK", data=data)
