+++ src/dashboard/apigateway/apigateway/apps/metrics/prometheus/__init__.py
@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 #
 # TencentBlueKing is pleased to support the open source community by making
 # 蓝鲸智云 - API 网关(BlueKing - APIGateway) available.
@@ -16,10 +15,3 @@
 # We undertake not to change the open source license (MIT license) applicable
 # to the current version of the project delivered to anyone in the future.
 #
-from django.urls import path
-
-from apigateway.apps.metrics import views
-
-urlpatterns = [
-    path("query_range/", views.QueryRangeAPIView.as_view(), name="metrics.query_range"),
-]
