+++ src/dashboard/apigateway/apigateway/tests/apis/web/metrics/test_views.py
@@ -17,11 +17,36 @@
 # to the current version of the project delivered to anyone in the future.
 #
 
+import pytest
 
-class TestQueryRangeAPIView:
+from apigateway.apis.web.metrics.views import MetricsSmartTimeRange
+
+
+class TestMetricsSmartTimeRange:
+    @pytest.mark.parametrize(
+        "time_range_minutes, expected",
+        [
+            (10, "1m"),
+            (59, "1m"),
+            (60, "1m"),
+            (300, "5m"),
+            (360, "5m"),
+            (720, "10m"),
+            (1440, "30m"),
+            (4320, "1h"),
+            (10080, "3h"),
+            (20000, "12h"),
+        ],
+    )
+    def test_get_recommended_step(self, time_range_minutes, expected):
+        smart_time_range = MetricsSmartTimeRange(time_range=time_range_minutes * 60)
+        assert smart_time_range.get_recommended_step() == expected
+
+
+class TestQueryRangeApi:
     def test_get(self, mocker, fake_stage, request_view):
         mocker.patch(
-            "apigateway.apps.metrics.views.DimensionMetricsFactory.create_dimension_metrics",
+            "apigateway.apis.web.metrics.views.DimensionMetricsFactory.create_dimension_metrics",
             return_value=mocker.Mock(query_range=mocker.Mock(return_value={"foo": "bar"})),
         )
