+++ src/dashboard/apigateway/apigateway/urls.py
@@ -60,7 +60,6 @@ urlpatterns = [
     path("backend/apis/<int:gateway_id>/ssl/", include("apigateway.apps.ssl_certificate.urls")),
     # apps: normal
     path("backend/apis/<int:gateway_id>/logs/", include("apigateway.apps.access_log.urls")),
-    path("backend/apis/<int:gateway_id>/metrics/", include("apigateway.apps.metrics.urls")),
     path("backend/apis/<int:gateway_id>/monitors/", include("apigateway.apps.monitor.urls")),
     path("backend/apis/<int:gateway_id>/audits/", include("apigateway.apps.audit.urls")),
     path("backend/apis/<int:gateway_id>/permissions/", include("apigateway.apps.permission.urls")),
@@ -84,15 +83,15 @@ urlpatterns = [
     path("backend/docs/esb/", include("apigateway.apps.docs.esb.urls")),
     path("backend/docs/feature/", include("apigateway.apps.docs.feature.urls")),
     path("backend/docs/feedback/", include("apigateway.apps.docs.feedback.urls")),
-
     # refactoring begin ------
     path("backend/apis/<int:gateway_id>/tests/", include("apigateway.apis.web.api_test.urls")),
     path("backend/gateways/<int:gateway_id>/tests/", include("apigateway.apis.web.api_test.urls")),
-
     # delete it later after frontend changed the url
     path("backend/apis/<int:gateway_id>/labels/", include("apigateway.apis.web.label.urls")),
     path("backend/gateways/<int:gateway_id>/labels/", include("apigateway.apis.web.label.urls")),
     path("backend/feature/", include("apigateway.apis.web.feature.urls")),
+    path("backend/apis/<int:gateway_id>/metrics/", include("apigateway.apis.web.metrics.urls")),
+    path("backend/gateways/<int:gateway_id>/metrics/", include("apigateway.apis.web.metrics.urls")),
     # refactoring end ------
 ]
